Bug 1849146 - Add tab and history search smoke tests
